[SYCL][COMPAT] Split of util headers into math and util
